+++ app/src/main/java/com/tovo/eat/ui/account/favorites/favkitchen/FavKitchenAdapter.java
@@ -133,7 +133,7 @@ public class FavKitchenAdapter extends RecyclerView.Adapter<BaseViewHolder> impl
 
         void offersItemClick(CouponListResponse.Result offers);
 
-        void onItemClickData(Integer kitchenId);
+        void onItemClickData(Long kitchenId);
 
         void animateView(View view);
 
@@ -141,7 +141,7 @@ public class FavKitchenAdapter extends RecyclerView.Adapter<BaseViewHolder> impl
 
         void removeDishFavourite(Integer favId);
 
-        void addFav(Integer id, String fav);
+        void addFav(Long id, String fav);
     }
 
     public class EmptyViewHolder extends BaseViewHolder {
@@ -191,12 +191,12 @@ public class FavKitchenAdapter extends RecyclerView.Adapter<BaseViewHolder> impl
 
 
         @Override
-        public void onItemClick(Integer id) {
+        public void onItemClick(Long id) {
             mLiveProductsAdapterListener.onItemClickData(id);
         }
 
         @Override
-        public void addFavourites(Integer id, String fav) {
+        public void addFavourites(Long id, String fav) {
             mLiveProductsAdapterListener.addFav(id, fav);
         }
