+++ app/src/main/java/com/tovo/eat/ui/kitchendetails/FavTodaysMenuItemViewModel.java
@@ -35,7 +35,7 @@ public class FavTodaysMenuItemViewModel {
     public final ObservableField<String> sQuantity = new ObservableField<>();
     public final ObservableField<Integer> quantity = new ObservableField<>();
     public final ObservableField<Integer> product_id = new ObservableField<>();
-    public final ObservableField<Integer> makeit_userid = new ObservableField<>();
+    public final ObservableField<Long> makeit_userid = new ObservableField<>();
     public final ObservableBoolean isFavourite = new ObservableBoolean();
     public final ObservableBoolean isVeg = new ObservableBoolean();
     public final ObservableBoolean isFavouriteMenu = new ObservableBoolean();
@@ -484,12 +484,12 @@ public class FavTodaysMenuItemViewModel {
 
         void refresh();
 
-        Integer getEatId();
+        Long getEatId();
 
         void showToast(String msg);
 
 
-        void otherKitchenDish(Integer makeitId, Integer productId, Integer quantity, Integer price);
+        void otherKitchenDish(Long makeitId, Integer productId, Integer quantity, Integer price);
     }
 
 }
